requirements.txt
Django==5.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
mysqlclient==2.1.1